LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-zh-Hant-HK-expected.txt

FAIL navigator.language should be zh-HK. Was zh-TW.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
